Don't reset game when joining from lobby.
When a game is active with, say "potato" as its ID and a new player joins from
the lobby by writing "potato", this resets the ongoing game.

I have added a `create_new` boolean to the corresponding request and set it
differently when it's used from the lobby (`create_new: false`) and when it's
used from the board (`create_new: true`).